+++ docker/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 All notable changes to the docker containers will be documented in this file.
 
+### 2026-08-21
+- Added support for scan sub-rule IDs in packaged scan `IGNORE` and `OUTOFSCOPE` configuration (Issue 9394).
+
 ### 2026-07-06
 - Default to the Client Spider when -j (the modern option) specified.